+++ src/modules/admin/widgets/grids/TransformationGridView.php
@@ -15,8 +15,8 @@ use yii\data\ArrayDataProvider;
 use yii\helpers\Url;
 
 /**
+ * @extends GridView<Transformation>
  * @property ActiveDataProvider|ArrayDataProvider|null $dataProvider
- * @method Transformation getModel()
  */
 class TransformationGridView extends GridView
 {